extract project executor
